+++ src/config.rs
@@ -34,11 +34,15 @@ pub struct RootsConfig {
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
 pub struct BackupConfig {
     pub path: String,
+    #[serde(rename = "ignoredGames")]
+    pub ignored_games: Option<std::collections::HashSet<String>>,
 }
 
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
 pub struct RestoreConfig {
     pub path: String,
+    #[serde(rename = "ignoredGames")]
+    pub ignored_games: Option<std::collections::HashSet<String>>,
 }
 
 impl Default for ManifestConfig {
@@ -54,6 +58,7 @@ impl Default for BackupConfig {
     fn default() -> Self {
         Self {
             path: default_backup_dir(),
+            ignored_games: None,
         }
     }
 }
@@ -62,6 +67,7 @@ impl Default for RestoreConfig {
     fn default() -> Self {
         Self {
             path: default_backup_dir(),
+            ignored_games: None,
         }
     }
 }
@@ -148,6 +154,64 @@ impl Config {
                 });
             }
             checked.insert(path);
+        }
+    }
+
+    pub fn is_game_enabled_for_backup(&self, name: &str) -> bool {
+        match &self.backup.ignored_games {
+            None => true,
+            Some(ignored) => !ignored.contains(name),
+        }
+    }
+
+    pub fn enable_game_for_backup(&mut self, name: &str) {
+        match &mut self.backup.ignored_games {
+            None => {}
+            Some(games) => {
+                games.remove(name);
+            }
+        }
+    }
+
+    pub fn disable_game_for_backup(&mut self, name: &str) {
+        match &mut self.backup.ignored_games {
+            None => {
+                let mut set = std::collections::HashSet::new();
+                set.insert(name.to_owned());
+                self.backup.ignored_games = Some(set);
+            }
+            Some(games) => {
+                games.insert(name.to_owned());
+            }
+        }
+    }
+
+    pub fn is_game_enabled_for_restore(&self, name: &str) -> bool {
+        match &self.restore.ignored_games {
+            None => true,
+            Some(ignored) => !ignored.contains(name),
+        }
+    }
+
+    pub fn enable_game_for_restore(&mut self, name: &str) {
+        match &mut self.restore.ignored_games {
+            None => {}
+            Some(games) => {
+                games.remove(name);
+            }
+        }
+    }
+
+    pub fn disable_game_for_restore(&mut self, name: &str) {
+        match &mut self.restore.ignored_games {
+            None => {
+                let mut set = std::collections::HashSet::new();
+                set.insert(name.to_owned());
+                self.restore.ignored_games = Some(set);
+            }
+            Some(games) => {
+                games.insert(name.to_owned());
+            }
         }
     }
 }